refactor toolbar, extracting code into separate components
